+++ Algo.h
@@ -12,12 +12,12 @@ using std::vector;
 class Algo
 {
 public:
-    vector<vector<int>> distances;
+    vector<vector<double>> distances;
     string name;
-    Algo(std::vector<std::vector<int>> distances, std::string name, std::default_random_engine rng);
+    Algo(std::vector<std::vector<double>> distances, std::string name, std::default_random_engine rng);
     std::default_random_engine rng;
     virtual Result solve() = 0;
-    int calculate_cost(vector<int> solution);
+    double calculate_cost(vector<int> solution);
     string get_name();
 };
